% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management_functions.R
\name{FL}
\alias{FL}
\title{FL}
\usage{
Determine the filtering loss (FL) for count
data based on removed features (adjusted from katiasmirn/PERfect/FL_J.R)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{new_counts}{Count matrix with samples as rows and features as counts after being partially filtered by SCRuB method}

\item{removed}{Vector of features to be removed as contaminants}
}
\value{
Data frame with features as row names and associated filtering loss value
}
\description{
FL
}
