% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline1.R
\name{pipeline1}
\alias{pipeline1}
\title{pipeline1}
\usage{
pipeline1(counts, meta, control_order = NA, seed = 42)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{Data frame with columns is_control, sample_type, and (optional) sample_well}

\item{control_order}{Vector ordering run of sample_type controls, default NA}

\item{seed}{Random seed}
}
\value{
List object with original matrix, decontaminated matrix, and corresponding
filtering loss (FL) statistics
}
\description{
Using the SCRuB method, this pipeline is best used when the user wants
to closely characterize the composition of the data prior to contamination.
}
\concept{pipeline1}
