% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline2.R
\name{step1}
\alias{step1}
\title{step1}
\usage{
step1(counts, meta)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{Matrix with columns is_control, sample_type, and batch}
}
\value{
Vector of features tagged as contaminants
}
\description{
Run step 1 of pipeline 2 to identify features that are differentially abundant
between batches
}
\seealso{
Other pipeline2: 
\code{\link{pipeline2}()},
\code{\link{step2}()},
\code{\link{step3}()},
\code{\link{step4}()}
}
\concept{pipeline2}
