% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline2.R
\name{step2}
\alias{step2}
\title{step2}
\usage{
Run step 2 of pipeline 2 to identify features that are expressed higher in negative
controls and lower in samples
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{Matrix with columns is_control, sample_type, and batch}

\item{threshold}{Threshold value for prevalence method of decontam}
}
\value{
Vector of features tagged as contaminants
}
\description{
step2
}
\seealso{
Other pipeline2: 
\code{\link{pipeline2}()},
\code{\link{step1}()},
\code{\link{step3}()},
\code{\link{step4}()}
}
\concept{pipeline2}
