% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline2.R
\name{step3}
\alias{step3}
\title{step3}
\usage{
Run step 3 of pipeline 2 to identify features with different abundance in technical
replicates across batches
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{technical_replicates}{Matrix identifying technical replicates across batches with batch as column and rows matching replicates}

\item{meta}{Matrix with columns is_control, sample_type, and batch}
}
\value{
Vector of features tagged as contaminants
}
\description{
step3
}
\seealso{
Other pipeline2: 
\code{\link{pipeline2}()},
\code{\link{step1}()},
\code{\link{step2}()},
\code{\link{step4}()}
}
\concept{pipeline2}
