% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline2.R
\name{step4}
\alias{step4}
\title{step4}
\usage{
step4(counts, blocklist)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{blocklist}{List of known previously identified contaminant features}

\item{meta}{Matrix with columns is_control, sample_type, and batch}
}
\value{
Vector of features tagged as contaminants
}
\description{
Run step 4 of pipeline 2 to identify features previously identified as contaminants
based on blocklist. Requires blocklist items of the lowest taxonomy.
}
\seealso{
Other pipeline2: 
\code{\link{pipeline2}()},
\code{\link{step1}()},
\code{\link{step2}()},
\code{\link{step3}()}
}
\concept{pipeline2}
