% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well2well.R
\name{well2well}
\alias{well2well}
\title{well2well}
\usage{
Determine the potential impact of spatial well to well relationships
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{Metadata with column one 'is_control' indicating TRUE if control, FALSE if not and 'sample_type' with sample name}

\item{control_name}{String indicating name of the control samples. Set to 'Control' if unspecified.}

\item{seed}{Random seed}
}
\description{
well2well
}
