% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management_functions.R
\name{wrap_phyloseq}
\alias{wrap_phyloseq}
\title{wrap_phyloseq}
\usage{
Wrap count matrix and metadata into a phyloseq object
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{dataframe with columns is_control, sample_type}
}
\value{
List containing counts and metadata data frames for input into pipeline functions
}
\description{
wrap_phyloseq
}
\seealso{
Other management: 
\code{\link{unwrap_phyloseq}()}
}
\concept{management}
