% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_pipeline}
\alias{visualize_pipeline}
\title{visualize_pipeline
Visualize results from the pipelines within the package}
\usage{
visualize_pipeline(pipeline_output, interactive = FALSE)
}
\arguments{
\item{pipeline_output}{Output of pipeline object}

\item{interactive}{TRUE if user wants interactive plot output}
}
\value{
Visualizations relating to pipeline object
}
\description{
visualize_pipeline
Visualize results from the pipelines within the package
}
